+++ src/init/point.c
@@ -3,10 +3,10 @@
 /*                                                        :::      ::::::::   */
 /*   point.c                                            :+:      :+:    :+:   */
 /*                                                    +:+ +:+         +:+     */
-/*   By: drenassi <marvin@42.fr>                    +#+  +:+       +#+        */
+/*   By: nsalles <nsalles@student.42perpignan.fr    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2024/02/17 13:31:47 by nsalles           #+#    #+#             */
-/*   Updated: 2024/02/17 20:22:21 by drenassi         ###   ########.fr       */
+/*   Updated: 2024/02/18 02:12:14 by nsalles          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -36,7 +36,7 @@ void	destroy_point(t_point *point)
 /*
  *	Create, set the values and return point struct.
 */
-t_point	*set_point(float x, float y, float z)
+t_point	*set_point(double x, double y, double z)
 {
 	t_point	*point;
